refactor: #124 rename CharacterItemCreateForm to more appropriate name ItemEditForm
