+++ resources/js/Components/Form/index.ts
@@ -1,6 +1,6 @@
 export {default as CharacterCreateForm} from './CharacterCreateForm'
 export {default as CharacterImportForm} from './CharacterImportForm'
-export {default as CharacterItemCreateForm} from './CharacterItemCreateForm'
 export {default as DmEntryCreateForm} from './DmEntryCreateForm'
+export {default as ItemEditForm} from './ItemEditForm'
 export {default as LoginForm} from './LoginForm'
 export {default as RegistrationForm} from './RegistrationForm'
